requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asn1crypto=1.3.0=py37_0
blas=1.0=mkl
boto=2.49.0=py37_0
boto3=1.11.0=py_0
botocore=1.14.0=py_0
bz2file=0.98=py37_1
ca-certificates=2019.11.27=0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
cryptography=2.8=py37h7a1dbc1_0
docutils=0.15.2=py37_0
flask=1.1.1=py_0
flask-wtf=0.14.2=py37_1
gensim=3.8.0=py37hf9181ef_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.8=py37_0
intel-openmp=2019.4=245
itsdangerous=1.1.0=py37_0
jinja2=2.10.3=py_0
jmespath=0.9.4=py_0
joblib=0.14.1=py_0
markupsafe=1.1.1=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
numpy=1.18.1=pypi_0
numpy-base=1.18.1=py37hc3f5095_0
openssl=1.1.1d=he774522_3
pip=19.3.1=py37_0
pycparser=2.19=py37_0
pyopenssl=19.1.0=py37_0
pysocks=1.7.1=py37_0
python=3.7.6=h60c2a47_2
python-crfsuite=0.9.6=pypi_0
python-dateutil=2.8.1=py_0
pyvi=0.0.9.7=pypi_0
requests=2.22.0=py37_1
s3transfer=0.3.0=py37_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=pypi_0
setuptools=44.0.0=py37_0
six=1.14.0=pypi_0
sklearn-crfsuite=0.3.6=pypi_0
smart_open=1.9.0=py_0
sqlite=3.30.1=he774522_0
tabulate=0.8.6=pypi_0
tqdm=4.41.1=py_0
urllib3=1.25.7=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wtforms=2.2.1=py37_0
